Created authorization using GATES

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\QuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -59,6 +60,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(Gate::denies('update-question', $question)){
+            abort(403, "Access denied!");
+        }
         return view("questions.edit", compact("question"));
     }
 
@@ -67,6 +71,10 @@ class QuestionsController extends Controller
      */
     public function update(QuestionRequest $request, Question $question)
     {
+        if(Gate::denies('update-question', $question)){
+            abort(403, "Access denied!");
+        }
+
         $question->update([
             'title' => $request->title,
             'body' => $request->body
@@ -80,6 +88,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question', $question)){
+            abort(403, "Access denied!");
+        }
+
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted!');
     }
